[WFLY-11754] Remove camel case from package names in ws testsuite.
